Prevent some cases of bad coverage for GAP files
If a file in the package being tested has the same name as a file in GAP, or
in another GAP package that happens to be loaded or somehow else in use, then
this can lead to coverage data for that file being mixed up with the package
we are actually testing.

Thus, remove any coverage data that comes from inside $HOME/gap/, i.e., not
from the package being tested.